Começando a adicionar versionamento na API com VO e terminando todos os repositories e services

+++ servidor/rest-pdv-padaria/src/main/java/br/com/leuxam/services/CompraEstoqueService.java
@@ -25,23 +25,30 @@ public class CompraEstoqueService {
 		return repository.findAll();
 	}
 	
+	public List<CompraEstoque> findAllWithProduto(Long id){
+		return repository.findAllWithProduto(id);
+	}
+	
+	public List<CompraEstoque> findAllWithCompra(Long id){
+		return repository.findAllWithCompras(id);
+	}
+	
 	public CompraEstoque create(CompraEstoque compraEstoque) {
 		return repository.save(compraEstoque);
 	}
-	
+
 	public CompraEstoque update(CompraEstoque compraEstoque) {
-		var entity = repository.findById(compraEstoque.getCompras().getId())
-				.orElseThrow(() -> new ResourceNotFoundException("Não existe nenhuma compra com esse ID"));
-		entity.setEstoque(compraEstoque.getEstoque());
-		entity.setPreco(compraEstoque.getPreco());
+		var entity = repository.findByIdProdutoAndCompras(compraEstoque.getEstoque().getId(), compraEstoque.getCompras().getId());
+		if(entity == null) throw new ResourceNotFoundException("Não encontrado o relatório com os ID's informados");
 		entity.setQuantidade(compraEstoque.getQuantidade());
+		entity.setPreco(compraEstoque.getPreco());
 		return repository.save(entity);
 	}
 	
-	public void delete(Long id) {
-		var entity = repository.findById(id)
-				.orElseThrow(() -> new ResourceNotFoundException("Não existe nenhuma compra com esse ID"));
+	public void delete(Long idProduto, Long idCompra) {
+		if(idProduto == 0 || idCompra == 0) throw new ResourceNotFoundException("Id do Produto ou da Compra não estão sendo informados");
+		var entity = repository.findByIdProdutoAndCompras(idProduto, idCompra);
+		if(entity == null) throw new ResourceNotFoundException("Não encontrado o relatório com os ID's informados");
 		repository.delete(entity);
 	}
-	
 }
